% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{A.make.stable}
\alias{A.make.stable}
\alias{optimal.PQ}
\alias{psi.fn}
\alias{orthonormalization_complex}
\title{AR(p) nearest stable system using succesive convex approximation}
\usage{
A.make.stable(a.unstable, ...)

optimal.PQ(A, n, p)

psi.fn(lambda, d, enumerator, alpha = 0.999)

orthonormalization_complex(u = NULL, basis = TRUE, norm = TRUE)
}
\arguments{
\item{a.unstable}{The unstable AR parameters in matrix form (n*np)}

\item{alpha}{The size of the Dikin Elipsoid}

\item{init.method}{The different initialization methods 'EV.ref1', 'EV.ref2', 'EV.scale', 'A.scale'}

\item{epsilon}{Convergence criterion}

\item{makestable.tol}{The minimal length of step $h$}
}
\value{
A list with the following elements
\describe{
  \item{a}{The nearest stable AR polynomial}
  \item{init}{The AR polynomial used for the initialization}
  \item{lam}{The step size at each iteration}
  \item{norm.step}{The norm of the matrix which has been added at each iteration step}
  \item{steps}{The number of steps needed until a stable \eqn{A} has been reached}
}
}
\section{Functions}{
\itemize{
\item \code{optimal.PQ}: function that calculates optimal P and Q

\item \code{psi.fn}: objective function for stats::uniroot

\item \code{orthonormalization_complex}: Orthonomalization of a matrix
}}

