% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{array.reorder}
\alias{array.reorder}
\title{is base::aperm doing the same?
array.reorder(matrix.to.array(cbind(diag(1, 3), -a)), a0=TRUE)-aperm(matrix.to.array(cbind(diag(1, 3), -a)), c(3, 1, 2))}
\usage{
array.reorder(a, a0 = FALSE, append.a0.only = FALSE)
}
\description{
convert a 3-dim array with dims [n,n,lag] into array with dims [lag,n,n]
a0=FALSE: add a0=I_n
only used to convert m$a into input of dse::ARMA
}
