% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{Blaschke.transformation}
\alias{Blaschke.transformation}
\title{Stabilize estimator of A using the idea of Blaschke.}
\usage{
Blaschke.transformation(a.unstable, sigma = NULL)
}
\arguments{
\item{a.unstable}{The unstable AR system parameters in matrix form (n*np)}

\item{sigma}{The variance matrix \eqn{\Sigma} of the innovations.}
}
\value{
a The stable polynomial

sigma The corresponding sigma
}
\description{
We are assuming that a(z) has different zeros. The idea is to reflect the roots which are outside the unit circle.
}
