% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator.R
\name{EM}
\alias{EM}
\alias{EM.algorithm}
\title{The EM algorithm for state space models}
\usage{
EM(dat.mf, p, n.init, iter.max, tol = 1e-04)

EM.algorithm(y, a, b, nf, N, iter.max, tol)
}
\arguments{
\item{dat.mf}{The mixed-frequency data (T*n matrix), where the first nf components are the fast ones and the last ns components are the slow.}

\item{p}{Order of the AR polynomial}

\item{n.init}{Number of initial values, randomly generated}

\item{iter.max}{Maximum number of iterations}

\item{tol}{Convergence tolerance for algorithm}
}
\value{
MLE parameters a.hat and b.hat
}
\description{
The EM algorithm for state space models

Fast Rcpp-EM Algorithm (EM0002)
}
