% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_sigma.R
\name{est.sigma}
\alias{est.sigma}
\title{Estimation of the innovation covariance matrix \eqn{\Sigma}}
\usage{
est.sigma(dat.mf, a)
}
\arguments{
\item{dat.mf}{The mixed-frequency data (T*n matrix)}

\item{a}{The estimated AR coefficients (n*np matrix)}
}
\value{
sigma The non-projected estimator of the innovation matrix \eqn{\Sigma}
}
\description{
Given the AR coefficients and the data,
an estimate for the innovation covariance matrix \eqn{\Sigma} is computed.
}
\seealso{
sigma.hat
}
