% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Estep0002}
\alias{Estep0002}
\title{The Expectation step in the EM Algorithm}
\usage{
Estep0002(Y, A, B, C, H, S, a1, P1, N, nf)
}
\description{
Compute the sufficient statistics in the EM Algorithm for state space models. Ignore the starting values.
}
