% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Estep0003}
\alias{Estep0003}
\title{The Expectation step in the EM Algorithm}
\usage{
Estep0003(Y, A, B, C, H, S, a1, P1, N, nf)
}
\description{
Compute the sufficient statistics in the EM Algorithm for state space models. Including starting values x1 with mean mu1 and covariance V1. See Shumway and Stoffer 1982.
}
