% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pars.R
\name{gen_random_numbers}
\alias{gen_random_numbers}
\title{Generate random numbers}
\usage{
gen_random_numbers(r.roots = 2, c.roots = 0, ...)
}
\arguments{
\item{r.roots}{The number of complex numbers, which should be generated}

\item{c.roots}{The number of complex numbers, which should be generated}

\item{real.r}{A vector of ranges of the moduli for the real numbers}

\item{complex.r}{A vector of ranges of the moduli for the complex numbers}

\item{ran}{A vector indicating the range, used if r.r an c.r are not specified}
}
\value{
Random real and complex random numbers
}
\description{
Generate random numbers
}
