% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{initialization}
\alias{initialization}
\title{Different initialization algorithms for the stabilization algorithm for the AR polynomial}
\usage{
initialization(a.unstable, init.method = "EV.ref1", ...)
}
\arguments{
\item{a.unstable}{The unstable AR system parameters in matrix form (n*np)}

\item{init.method}{The name of the initialization method 'EV.ref1', 'EV.ref2', 'EV.scale', 'A.scale'}
}
\value{
a.stable The calculated stable AR polynomial
}
\description{
The following methods are available:
}
