% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator.R
\name{IVL}
\alias{IVL}
\alias{AIC_IVL}
\title{Instrumental Variable (IVL) estimator for MF-data}
\usage{
IVL(dat.mf, p, k = NULL, upper = NULL)

AIC_IVL(dat.mf, p, n, nf, N, grid, upper = NULL)
}
\arguments{
\item{dat.mf}{The mixed-frequency data (T*n matrix), where the first nf components are the fast ones and the last ns components are the slow.}

\item{p}{Order of the AR polynomial}

\item{k}{A predefined k. If it is NULL, it will be estimated}

\item{upper}{The upper bound for the k estimation. If it is NULL, then log(n_s)^1.5 will be used where n_s is the amount of observed slow components.}
}
\value{
a.hat The estimator of the AR polynomial

k The estimated lag for the
}
\description{
The method is described in L. Koelbl, M. Deistler (2018): A New Approach for Estimating VAR Systems in the Mixed-Frequency Case (Statistical Papers)
The main idea of this new procedure is to project the slow components on the present and past fast ones in order to create instrumental variables.
}
\section{Functions}{
\itemize{
\item \code{AIC_IVL}: estimation of the k parameter
}}

