% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kfilter}
\alias{kfilter}
\title{The Kalman filter equations for state space systems}
\usage{
kfilter(Y, A, B0, C, H, S, a1, P1, N, nf)
}
\arguments{
\item{Y}{The data matrix}

\item{A}{The transiation matrix}

\item{C}{The observation matrix}

\item{H}{The observation error covariance matrix}

\item{S}{The covariance between observation and transition error}

\item{a1}{The mean of state x1}

\item{P1}{The covariance of state x1}

\item{N}{The sampling frequency of the slow component}

\item{nf}{The number of fast components}

\item{B}{The matrix corresponding to the transition error, i.e. \eqn{\Sigma=BB'}}
}
\value{
A list with the usual Kalman filter variables.
}
\description{
The Kalman filter equations for state space systems
}
