% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pars.R
\name{modgen.a}
\alias{modgen.a}
\title{Randomly generate an AR polynomial \eqn{a(z)} in matrix form}
\usage{
modgen.a(n, p, ...)
}
\arguments{
\item{n}{The dimension of the process.}

\item{p}{The order of the autoregressive system.}

\item{complex.roots}{The number of complex roots of the polynomial \eqn{a(z)}.}
}
\value{
a
}
\description{
Randomly generate an AR polynomial \eqn{a(z)} in matrix form
}
\seealso{
gen_random_numbers
}
