% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pars.R
\name{modgen.b}
\alias{modgen.b}
\title{Generate a n*q matrix where columns have length 1}
\usage{
modgen.b(n, ...)
}
\arguments{
\item{n}{The dimension of the process.}

\item{q}{The rank of sigma matrix.}
}
\value{
Matrix b of dimension n*q
}
\description{
Generate a n*q matrix where columns have length 1
}
