% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pars.R
\name{modgen}
\alias{modgen}
\title{Generate randomly a model \eqn{a(z)}, b, sigma}
\usage{
modgen(n, p, ...)
}
\arguments{
\item{n}{Dimension of the process.}

\item{p}{Order of the autoregressive polynomial \eqn{a(z)}.}
}
\value{
a The AR polynomial in matrix form.

b

Sigma The variance matrix of the innovations.
}
\description{
Generate randomly a model \eqn{a(z)}, b, sigma
}
