% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{proj.sigma}
\alias{proj.sigma}
\title{Projection of a symmetric matrix on the space of positive semi-definite matrices.}
\usage{
\method{proj}{sigma}(sigma, q = NULL, epsilon = 1e-07)
}
\arguments{
\item{sigma}{A symmetric matrix of dimension n*n.}

\item{q}{q is the desired rank of sigma.}

\item{epsilon}{Negative eigenvalues will be set to epsilon.}
}
\value{
sigma.proj The projected estimator of the innovation matrix.
}
\description{
Projection of a symmetric matrix on the space of positive semi-definite matrices.
}
