% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_sigma.R
\name{sigma.hat}
\alias{sigma.hat}
\title{Estimation and projection of the innovation covariance matrix \eqn{\Sigma}}
\usage{
\method{sigma}{hat}(dat.mf, a, ...)
}
\arguments{
\item{dat.mf}{The mixed-frequency data (T*n matrix)}

\item{a}{The estimated AR coefficients (n*np matrix)}

\item{q}{The desired rank of the innovation matrix \eqn{\Sigma} (n>=q>0)}
}
\value{
sigma The projected estimator of the innovation matrix \eqn{\Sigma}

sigma.0 The non-projected estimator of the innovation matrix \eqn{\Sigma}
}
\description{
Estimation of the innovation covariance matrix \eqn{\Sigma}  and
projection onto the space of positive-semidefinite matrices.
The estimation procedure and the projection is described in
L. Koelbl, A. Braumann, E. Felsenstein, M. Deistler (2016): Estimation of VAR Systems from Mixed Frequency Data: The Stock and the Flow Case, Advances in Econometrics, 35, pp. 43-73, DOI: 10.1108/S0731905320150000035002
}
\details{
If q is not desired, it will be set automatically to n, i.e. dimension of the innovations. Only the stock case is supported up to now.
}
