% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ar.to.ss}
\alias{ar.to.ss}
\title{state space representation of varp system}
\usage{
ar.to.ss(a, b, R_ii = NULL, scheme = NULL, nf = NULL)
}
\arguments{
\item{R_ii}{The diagonal element of the measurement covariance matrix \eqn{R}}

\item{scheme}{Weights for flow data}
}
\description{
state space representation of varp system
}
