% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator.R
\name{XYW}
\alias{XYW}
\title{XYW - Extended Yule Walker estimator for MF-data}
\usage{
XYW(dat.mf, p, k = NULL)
}
\arguments{
\item{dat.mf}{The mixed-frequency data (T*n matrix), where the first nf components are the fast ones and the last ns components are the slow.}

\item{p}{Order of the AR polynomial}

\item{k}{A predefined k, which specifies the number columsn of \eqn{Z_0}. (TODO) If it is NULL, it will be estimated chosen such that the condition number of \eqn{Z_0} is minimal (the maximal number of columns of \eqn{Z_0}, denote by kmax, kmax = some number...)}
}
\value{
a.hat The estimator of the AR polynomial
}
\description{
An estimator for the system-parameters of a VAR(p) process based on Extended-Yule Walker equations
}
